Commit data prefab experimental result
Data prefab assign to blob asset can be overwrite after next creation. Prefab can be created and the conversion will take place automatically. This means that the prefab can be loaded and created anytime that is necessary. Making a system to watch for managed reference type of the tag and remove entity if query found. This system run after conversion so there won't be any issue.