Mainly add for show hide level objects
The level objects that are converted from sub scene has SceneTag component which can be used to check
Merge in waypoint and flowfield modules and refine. This is WIP
Create assist to show flow field so that later flow field can be corrected with some visual aid. This is also WIP